+++ react/src/component/label.js
