+++ react/src/component/panel.js
