+++ react/src/context.js
@@ -10,8 +10,7 @@ export const ContextSimulationListPromise = React.createContext();
 export const ContextSimulationInfoPromise = React.createContext();
 
 export const ContextAppName = React.createContext();
-export const ContextAppViewBuilder = React.createContext();
-export const ContextAppInfo = React.createContext();
+export const ContextLayouts = React.createContext();
 
 export const ContextRelativeRouterHelper = React.createContext();
 export const ContextRelativeFormController = React.createContext();
