+++ react/src/data/dependency.js
@@ -1,5 +1,3 @@
-import { useSelector, useDispatch, useStore } from 'react-redux';
-
 export class Dependency {
     constructor(dependencyString) {
         let { modelName, fieldName } = this.mapDependencyNameToParts(dependencyString);
