+++ react/src/hook/legacy.js
@@ -16,18 +16,3 @@
     }, [lock, ...(reqs || [])])
     return () => updateLock(false);
 }
-
-export function useSetup(shouldRun, callback) {
-    const [hasSetup, updateHasSetup] = useState(false);
-    const [callbackStarted] = useState({value: false});
-    const finish = () => {
-        updateHasSetup(true);
-    }
-    useEffect(() => {
-        if(shouldRun && !hasSetup && !callbackStarted.value) {
-            callbackStarted.value = true;
-            callback(finish);
-        }
-    });
-    return hasSetup;
-}
