+++ react/src/hook/string.js
