+++ react/src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { BrowserRouter } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.css';
-import { RootRouter } from './app/root';
-import AppRoot from './app/app';
+import { RootRouter } from './component/root';
+import { AppRoot } from './component/app';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
