+++ react/src/layout/navbar.js
@@ -1,10 +1,4 @@
-import { useContext, useState } from "react";
-import { Modal } from "react-bootstrap";
-import { useInterpolatedString } from "../dependency/dependency";
-import { ContextModels } from "../components/context";
-import { elementForLayoutName } from "./panel";
-
-export let NavBarModalButton = ({ schema }) => (view) => {
+/*export let NavBarModalButton = ({ schema }) => (view) => {
     let NavBarModalButtonComponent = (props) => {
         let { config } = view;
         let { modal } = config;
@@ -35,4 +29,4 @@ export let NavBarModalButton = ({ schema }) => (view) => {
             </Modal>
         )
     }
-}
+}*/
