+++ react/src/utility/object.js
